[260108] feat : 예약 시스템 API 구현 및 스케줄 엔티티 수정
- BookingController: 예약 신청 API 컨트롤러 구현
- ApiResponse: 공통 API 응답 Wrapper 클래스 추가
- BookingCreateRequest: 예약 신청 요청 DTO 수정
- BookingCreateResponse: 예약 성공 응답 DTO 추가
- Schedule: 스케줄 엔티티 필드 및 구조 변경
- ScheduleRepository: 스케줄 데이터 접근 로직 수정

+++ src/main/java/kr/java/pr1mary/repository/ScheduleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
     // 특정 선생님의 예약 가능한 스케줄 조회
     List<Schedule> findAllByUserIdAndIsBookedFalseOrderByStartTimeAsc(Long teacherId);
     // 특정 날짜 범위의 스케줄 조회(캘린더 뷰)
